+++ src/lib/sx1262/ace/src/rxdataframequeue.cpp
@@ -11,16 +11,9 @@
 
 GATAS::PostConstruct RxDataFrameQueue::postConstruct()
 {
-    dataQueue = xQueueCreate(2, sizeof(GATAS::DataFrame));
-    if (dataQueue == nullptr)
-    {
-        return GATAS::PostConstruct::XQUEUE_ERROR;
-    }
-
     // Task must be lower compared to any SX1262 task
     if (xTaskCreate(radioQueueTaskTrampoline, RxDataFrameQueue::NAME.cbegin(), configMINIMAL_STACK_SIZE + 1024, this, tskIDLE_PRIORITY + 2, &taskHandle) != pdPASS)
     {
-        vQueueDelete(dataQueue);
         return GATAS::PostConstruct::TASK_ERROR;
     }
 
@@ -32,11 +25,24 @@ void RxDataFrameQueue::start()
     getBus().subscribe(*this);
 };
 
+bool RxDataFrameQueue::push(GATAS::DataFrame &&frame)
+{
+    if (!dataQueue.push(etl::move(frame)))
+    {
+        statistics.queueFull += 1;
+        return false;
+    }
+
+    xTaskNotifyGive(taskHandle);
+    return true;
+}
+
 void RxDataFrameQueue::getData(etl::string_stream &stream, const etl::string_view path) const
 {
     (void)path;
     stream << "{";
-    stream << "\"totalOutgoing:k\":\"" << statistics.totalOutgoing << "\"";
+    stream << "\"totalQueued:k\":\"" << statistics.totalQueued << "\",";
+    stream << "\"queueFull:err\":\"" << statistics.queueFull << "\"";
     stream << "}";
 }
 
@@ -53,60 +59,62 @@ void RxDataFrameQueue::radioQueueTask(void *arg)
     GATAS::DataFrame rxFrame;
     while (true)
     {
-        if (xQueueReceive(dataQueue, &rxFrame, portMAX_DELAY) == pdPASS)
+        if (!dataQueue.pop(rxFrame))
         {
-            PoolReleaseGuard guard{getGlobalPool(), rxFrame.frame};
+            uint32_t notifyValue = 0;
+            xTaskNotifyWait(pdFALSE, ULONG_MAX, &notifyValue, portMAX_DELAY);
+            continue;
+        }
 
-            statistics.totalOutgoing += 1;
-            if (rxFrame.config->manchester)
+        statistics.totalQueued += 1;
+        if (rxFrame.config->manchester)
+        {
+            size_t byteLength = rxFrame.length >> 1;
+            if (auto errorFrame = static_cast<uint8_t *>(getGlobalPool().alloc(byteLength)))
             {
-                size_t byteLength = rxFrame.length >> 1;
-                if (auto errorFrame = static_cast<uint8_t *>(getGlobalPool().alloc(byteLength)))
-                {
-                    guard.disarm();
-                    auto msg = GATAS::RadioRxManchesterMsg{
-                        rxFrame.frame,
-                        errorFrame,
-                        byteLength,
-                        rxFrame.epochSeconds,
-                        rxFrame.frequency,
-                        rxFrame.config->dataSource(),
-                        rxFrame.rssidBm};
-
-                    manchesterDecodeInline(msg.frame, msg.error, rxFrame.length);
-
-                    // Handle multi protocol situations
-                    // auto ds = decideDataSource(rxFrame.config->dataSource(), msg.frame32(), rxFrame.length );
-                    // if (ds.dataSource < GATAS::DataSource::_TRANSPROTOCOLS)
-                    // {
-                    //     bitShift(
-                    //         msg.frame,
-                    //         msg.lengthBytes,
-                    //         ds.bitsToShift);
-
-                    //     bitShift(
-                    //         msg.error,
-                    //         msg.lengthBytes,
-                    //         ds.bitsToShift);
-
-                    //     msg.lengthBytes = ds.frameLength;
-                    //     msg.dataSource = ds.dataSource;
-                    // }
-
-                    getBus().receive(msg);
-                }
-            }
-            else
-            {
-                guard.disarm();
-                getBus().receive(GATAS::RadioRxMsg{
-                    rxFrame.frame,
-                    rxFrame.length,
+                auto msg = GATAS::RadioRxManchesterMsg{
+                    getGlobalPool(),
+                    rxFrame.frame.detach(),
+                    errorFrame,
+                    byteLength,
                     rxFrame.epochSeconds,
                     rxFrame.frequency,
                     rxFrame.config->dataSource(),
-                    rxFrame.rssidBm});
+                    rxFrame.rssidBm};
+
+                manchesterDecodeInline(msg.frame.get(), msg.error.get(), rxFrame.length);
+
+                // Handle multi protocol situations
+                // auto ds = decideDataSource(rxFrame.config->dataSource(), msg.frame32(), rxFrame.length );
+                // if (ds.dataSource < GATAS::DataSource::_TRANSPROTOCOLS)
+                // {
+                //     bitShift(
+                //         msg.frame,
+                //         msg.lengthBytes,
+                //         ds.bitsToShift);
+
+                //     bitShift(
+                //         msg.error,
+                //         msg.lengthBytes,
+                //         ds.bitsToShift);
+
+                //     msg.lengthBytes = ds.frameLength;
+                //     msg.dataSource = ds.dataSource;
+                // }
+
+                getBus().receive(msg);
             }
+        }
+        else
+        {
+            getBus().receive(GATAS::RadioRxMsg{
+                getGlobalPool(),
+                rxFrame.frame.detach(),
+                rxFrame.length,
+                rxFrame.epochSeconds,
+                rxFrame.frequency,
+                rxFrame.config->dataSource(),
+                rxFrame.rssidBm});
         }
     }
 }