+++ src/lib/wifiservice/ace/src/wifiservice.cpp
@@ -63,7 +63,9 @@ void WifiService::wifiTask(void *arg)
     WifiService *wifiService = (WifiService *)arg;
     while (true)
     {
-        uint32_t notifyValue = ulTaskNotifyTake(pdTRUE, TASK_DELAY_MS(1'000));
+        uint32_t notifyValue = 0;
+        xTaskNotifyWait(pdFALSE, ULONG_MAX, &notifyValue, TASK_DELAY_MS(1'000));
+
         if (notifyValue != 0)
         {
             if (notifyValue & TaskState::SHUTDOWN)